notes: split into RolesField + SubstitutionsField + SubstitutionRow
NotesEditSection becomes a thin coordinator (~150 lines). Substitution
reorder now invalidates the party query on failure so the UI rehydrates
from the server instead of drifting from a partial write.
